requirements.txt
aiogram>=3.0
jinja2>=3.0
pytest>=7.0
pytest-asyncio>=0.20
psutil>=5.9
python-dotenv>=0.20